refactor: :sparkles: Adicionando algumas mensagens nas rotas

+++ src/service/LivroService.js
@@ -7,7 +7,7 @@ class LivroService {
       include: [{
         model: Autor,
         required: true,
-        attributes: ['nome'],
+        attributes: ['id', 'nome'],
       }],
     });
     return resultado;
@@ -20,24 +20,49 @@ class LivroService {
 
     const resultado = await Livro.findByPk(id);
 
+    if (resultado == null) {
+      throw new Error('ID não existe no Banco!');
+    }
+
     return resultado;
   }
 
   async criarLivro(body) {
     if (!body.autor_id) {
-      throw new Error('Necessario id de um autor');
+      throw new Error('ID do autor necessario');
     }
+
+    const autorID = await Autor.findByPk(body.autor_id);
+
+    if (autorID == null) {
+      throw new Error('ID do autor não existe no Banco!');
+    }
+
     const resultado = await Livro.create(body);
 
     return resultado;
   }
 
-  atualizarLivro(id, body) {
-    return Livro.update(body, { where: { id } });
+  async atualizarLivro(id, body) {
+    if (!Object.keys(body).length) {
+      throw new Error('Corpo da requisicao vazio');
+    }
+    const resultado = await Livro.update(body, { where: { id } });
+    if (resultado !== null) {
+      const livro = await Livro.findByPk(id);
+      return livro;
+    }
+    return resultado;
   }
 
-  excluirLivro(id) {
-    return Livro.destroy({ where: { id } });
+  async excluirLivro(id) {
+    if (!id) {
+      throw new Error('ID necessario para a excluisão!');
+    }
+
+    const resultado = await Livro.destroy({ where: { id } });
+
+    return resultado;
   }
 }
